elo implementation

+++ GwApiCore/src/main/java/us/gameandwatching/gwapi/core/users/UserFields.java
@@ -5,6 +5,6 @@ public interface UserFields {
   String getUsername();
   String getPasswordHash();
   String getPasswordSalt();
-  int getElo();
+  double getElo();
   long getCreatedAt();
 }
